print digit.c test1 for problem with long int

+++ print_digit.c
@@ -11,9 +11,9 @@ int print_digit (va_list dgt)
 {
 	int len = 0;
 	long int num = va_arg(dgt, int);
-	int div = 1;
-	int tmp;
-	int res1;
+	long int div = 1;
+	long int tmp;
+	long int res1;
 	int i;
 
 	if (num < 0)
